LPD-8518 Move to right package
